FirebaseConfig.js is created and global variable firebaseConfig is exported and imported in school.insert.js
