+++ Config/FirebaseConfig.js
@@ -1,0 +1,12 @@
+var firebaseConfig = {
+    apiKey: "AIzaSyDlESaTQUPM3rRX_BBLLihLfVf3UEEfBmc",
+    authDomain: "admission-consultant-c9db1.firebaseapp.com",
+    databaseURL: "https://admission-consultant-c9db1.firebaseio.com",
+    projectId: "admission-consultant-c9db1",
+    storageBucket: "admission-consultant-c9db1.appspot.com",
+    messagingSenderId: "1006379150064",
+    appId: "1:1006379150064:web:2d4cb80711383179ef7035",
+    measurementId: "G-P82D3N1F4K"
+  };
+
+  export {firebaseConfig};
