+++ category/school/schoolInsert.html
@@ -224,7 +224,8 @@
     <ol id="List"></ol>
   </div>
   <script src="https://cdn.jsdelivr.net/npm/sweetalert2@9"></script>
-  <script src="schoolInsert.js"></script>
+  <script type="module" src="schoolInsert.js"></script>
+  <script type="module" src="../../Config/FirebaseConfig.js"></script>
 </body>
 
 </html>
